Changed all the longs to ints to make sure the cusparse calls will work.

+++ lanczos.cu
@@ -54,7 +54,7 @@ __global__ void arraysalloc(cuDoubleComplex** a, int n, int m){
 
 int main(){}
 
-void lanczos(const long num_Elem, const cuDoubleComplex* d_H_vals, const int* d_H_rows, const int* d_H_cols, const int dim, int max_Iter, const int num_Eig, const double conv_req){
+void lanczos(const int num_Elem, const cuDoubleComplex* d_H_vals, const int* d_H_rows, const int* d_H_cols, const int dim, int max_Iter, const int num_Eig, const double conv_req){
 
   cublasStatus_t linalgstat;
   //have to initialize the cuBLAS environment, or my program won't work! I could use this later to check for errors as well
